requirements.txt
customtkinter==5.2.2
SpeechRecognition==3.10.4
pyttsx3==2.90
pyaudio==0.2.14
openai==1.40.6
google-generativeai==0.7.2
